moving up main.py to root folder
