+++ .github/workflows/Build-And-Test.yml
@@ -51,4 +51,4 @@ jobs:
           SA_CLIENT_ID: ${{ secrets.SA_CLIENT_ID }}
           SA_CLIENT_X509_CERT_URL: ${{ secrets.SA_CLIENT_X509_CERT_URL }}
       run: |
-        python baseball-japan/main.py upload:no
+        python main.py upload:no
